Rename CustomAppConfig to be snake_case
